LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/script-supports-expected.txt


FAIL Type of HTMLScriptElement.supports is function assert_equals: expected "function" but got "undefined"
FAIL HTMLScriptElement.supports resurns true for 'classic' HTMLScriptElement.supports is not a function. (In 'HTMLScriptElement.supports('classic')', 'HTMLScriptElement.supports' is undefined)
FAIL HTMLScriptElement.supports resurns true for 'module' HTMLScriptElement.supports is not a function. (In 'HTMLScriptElement.supports('module')', 'HTMLScriptElement.supports' is undefined)
FAIL HTMLScriptElement.supports returns false for JavaScript MIME types HTMLScriptElement.supports is not a function. (In 'HTMLScriptElement.supports('application/ecmascript')', 'HTMLScriptElement.supports' is undefined)
FAIL HTMLScriptElement.supports returns false for unsupported types HTMLScriptElement.supports is not a function. (In 'HTMLScriptElement.supports('')', 'HTMLScriptElement.supports' is undefined)
